+++ client/routes.js
@@ -32,7 +32,6 @@ class Routes extends Component {
 
           <Route path="/about-alexa" component={AlexaInstruct} />
           <Route path="/privacy" component={Privacy} />
-          <Route path="/receipt" component={Receipt} />
 
           {isLoggedIn && (
             <Switch>
@@ -41,7 +40,7 @@ class Routes extends Component {
               <Route path="/home" component={DailyRecipe} />
               <Route path="/recipe" component={Recipe} />
               <Route path="/fridge" component={Fridge} />
-              <Route path="recipt" component={Recipt} />
+              <Route path="/receipt" component={Receipt} />
             </Switch>
           )}
           {/* Displays our Login component as a fallback */}
