+++ public/index.html
@@ -6,6 +6,7 @@
     <title class="headliner">Chef Remy</title>
     <link rel="stylesheet" href="/style.css" />
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+    <script src=”https://cdn.rawgit.com/naptha/tesseract.js/1.0.10/dist/tesseract.js”></script>
     <script defer src="/bundle.js"></script>
   </head>
   <body>
